+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- support importing component on top of a modified one and merging the changes by adding `--merge` flag to `bit import`
 - [#927](https://github.com/teambit/bit/issues/927) fix a case of link file (file that only requires other file) is part of the component
 - improve bit-diff to show the output with colors
 - fix bit-move of a directly imported dependency